pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ouragentteams"
version = "0.1.0"
description = "ourAgentTeams with local LLM Leader orchestration"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40",
    "ollama>=0.2",
    "chromadb>=0.5",
    "typer>=0.12",
    "rich>=13",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "tiktoken>=0.7",
    "sqlalchemy>=2",
    "pyyaml>=6",
    "aiofiles>=23",
]

[project.scripts]
ouragentteams = "src.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
